utils: Allow passing values converted to dbus types with signatures
So far it was not possible to pass a dbus.type() to to_variant()
function with a specified signature, as we had a table instead of the
real value.

Extract dbus type and compare with the signature for equality.
Additionally parse real value instead of the Lua table created by
dbus.type().