utils: Use 'av' signature in case of unknown first array element type
In case first element type was also an array are were going into
"default" case in switch statement, which was assuming that it is an
'ai' array. In fact we don't know the type of the array and it is not
easy to get it without recursively checking it's elements.

Use 'av' signature as default signature, so we don't need to check
table elements types, but send the table (or any other type) packed
into variant.